+++ authorization/src/main/java/org/grnet/status/authorizations/clients/KeycloakTokenClient.java
@@ -1,0 +1,20 @@
+package org.grnet.status.authorizations.clients;
+
+import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.MediaType;
+import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
+import org.grnet.status.authorizations.dtos.TokenResponse;
+
+@RegisterRestClient(configKey = "keycloak-token-client")
+@Path("/protocol/openid-connect/token")
+public interface KeycloakTokenClient {
+
+    @POST
+    @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
+    @Produces(MediaType.APPLICATION_JSON)
+    TokenResponse getToken(
+            @FormParam("grant_type") String grantType,
+            @FormParam("client_id") String clientId,
+            @FormParam("client_secret") String clientSecret
+    );
+}