+++ authorization/src/main/java/org/grnet/status/authorizations/dtos/GroupRequest.java
@@ -1,0 +1,15 @@
+package org.grnet.status.authorizations.dtos;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.List;
+import java.util.Map;
+
+public class GroupRequest {
+
+
+    @JsonProperty("name")
+    public String name;
+
+    @JsonProperty("attributes")
+    public Map<String, List<String>> attributes;
+}