+++ authorization/src/main/java/org/grnet/status/authorizations/dtos/TokenResponse.java
@@ -1,0 +1,8 @@
+package org.grnet.status.authorizations.dtos;
+
+public class TokenResponse {
+    public String access_token;
+    public long expires_in;
+    public long refresh_expires_in;
+    public String token_type;
+}